fix: remove non-output prints
